refactor: renamed classes and modules
renamed classes for better clarity and also made some classes singletone espectially the db engine connection procedure
