+++ campaign.py
@@ -1,11 +1,10 @@
-from os import environ, path
+from os import environ
 
 from dotenv import load_dotenv
-from sqlalchemy import create_engine
 
 from my_logger import MyLogger
-from MessageStatus import MessageStatus
-from fuelrod import campaign_queue as msg, fuelrod_api
+from fuelrod.fuelrod_api import MessageStatus
+from fuelrod.campaign_queue import Campaign
 
 load_dotenv(verbose=True)
 debug_db = environ.get('DEBUG_DB', False)
@@ -13,12 +12,7 @@ log_level = environ.get('LOG_LEVEL', 'INFO')
 
 logging = MyLogger()
 
-engine = create_engine("mysql://fuelrod:fuelrod@localhost/fuelrod",
-                       echo=debug_db,
-                       echo_pool=debug_db,
-                       hide_parameters=not debug_db)
-
-messages = msg.FuelrodQueue(db_engine=engine, my_logger=logging)
+messages = Campaign()
 
 campaigns = messages.load_unprocessed_campaigns(campaign_status=MessageStatus.COMPLETED)
