+++ my_logger.py
@@ -4,11 +4,25 @@ from colorlog import ColoredFormatter
 from os import path, mkdir, environ
 import datetime
 from dotenv import load_dotenv
+from sqlalchemy import create_engine
 
 load_dotenv(verbose=True)
 
 
-# coloredlogs.install()
+class MyDb:
+    db_engine = None
+    db_url = environ.get('DB_URL', 'mysql://fuelrod:fuelrod@localhost/fuelrod')
+    debug_db = environ.get('DEBUG_DB', False)
+
+    def __new__(cls, *args, **kwargs):
+        if cls.db_engine is None:
+            cls.db_engine = create_engine(url=cls.db_url,
+                                          echo=cls.debug_db,
+                                          echo_pool=cls.debug_db,
+                                          hide_parameters=not cls.debug_db)
+
+        return cls.db_engine
+
 
 class MyLogger:
     _logger = None
@@ -16,7 +30,6 @@ class MyLogger:
 
     def __new__(cls, *args, **kwargs):
         if cls._logger is None:
-            print("Logger new")
             cls._logger = super().__new__(cls, *args, **kwargs)
             cls._logger = logging.getLogger("fuelrod")
             cls._logger.setLevel(cls.log_level)
